LayoutTests/fast/scrolling/ios/overflow-scroll-overlap-3-expected.txt

Test that scrollable areas with text overlap are correctly targeted.

case 1:
case 2: Scrollable 2
case 3: Scrollable 3
case 4:
case 5: Scrollable 5
case 6: Scrollable 6
case 7:
case 8: Scrollable 8
